Add registration number, check highlight bug, fix bugs.
- accroding to the rules of MIIT, every website should show its
registration number on the front page.

- There is a conflict in principle between the export of myhighlight
and typora:
  - myhighlight set line number while loading the body of html, but
use selected style after the html is totally loaded.
  - the way to judge if html is from typora is to judge the class of
its body, but this needs to be after the load of html
  - So, if we check the class of body at the start of myhighlight, we
have to delay myhighlight after the load, which causes the load of
style wrong. We have to give up to combine them.

- Fix content bugs in vim, ohos_compile, and so on.