Add a button for nav list, and set styles
The nav list(named pandoc-menu) is an important part of code
blogs. However, it was always shown, no matter what you use, pc, mobile
or tablet. It makes chaos for mobile/tablet, because width of these
devices is insufficient. Then the list covers normal content.

In this commit, a button is used to control the open/close of nav list,
and button & list are in the same parent, who comtrols the position and
flow of them.

Although it's still ugly, we've moved a step for the better, isn't it?